+++ authpass/CHANGELOG.md
@@ -6,6 +6,7 @@
   * cmd/ctrl+c: Will only copy password if not in another text field.
   * ESC: Will blur the current text field. If inside the search bar, will cancel the search.
   * Linux: Workaround for pasting into text fields.
+* iOS: Allow copying of kdbx files from/to application docs dir with iTunes/Finder. #110
 
 # 1.7.4